# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GenBI API
  version: 0.1.0
  description: Generative BI Platform API
servers:
  - url: http://localhost:8000/api/v1
    description: Development
paths:
...
  /users:
    get:
      operationId: list_tenant_users
      summary: "The tenant's users — tenant admin or superuser"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User list
    post:
      operationId: create_tenant_user
      summary: "Create a user in the caller's tenant"
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Created
        "409":
          description: USER_EXISTS
  /users/{user_id}:
    patch:
      operationId: update_tenant_user
      summary: "Update email / roles / enable-disable"
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated
    delete:
      operationId: delete_tenant_user
      summary: "Hard delete — refused for the last active tenant admin"
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
        "422":
          description: LAST_TENANT_ADMIN
  /users/{user_id}/reset-password:
    post:
      operationId: admin_reset_user_password
      summary: "(Admin sets a new password for a tenant user"
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reset
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...